+++ src/app/core/type/index.ts
@@ -1,4 +1,5 @@
-export * from './feature.type';
-export * from './structure.type';
 export * from './app.type';
+export * from './feature.type';
+export * from './language.type';
+export * from './structure.type';